Added requirements.txt and updated code
I made the mistake of intitially developing this repo within
the same environment as the original Carla repo. I have now added
a requirements.txt file to ensure that newer updated versions of the
libs are being used. I have also gone through most of the code to
resolve the new issues that arose from the updated libraries.

So far the methods are working again, however, it is possible
that bugs now may exists that have not been caught. These will arise
with time and will likely get fixed. I also see some seemingly non-breaking
warnings, which should be resolved in the future as well.

+++ evaluation_layer/distances.py
@@ -28,7 +28,7 @@ def l0_distance(delta: np.ndarray) -> List[float]:
     num_feature_changes = np.sum(
         difference_mask,
         axis=1,
-        dtype=np.float,
+        dtype=np.float32,
     )
     distance = num_feature_changes.tolist()
     return distance
@@ -48,7 +48,7 @@ def l1_distance(delta: np.ndarray) -> List[float]:
     List[float]
     """
     absolute_difference = np.abs(delta)
-    distance = np.sum(absolute_difference, axis=1, dtype=np.float).tolist()
+    distance = np.sum(absolute_difference, axis=1, dtype=np.float32).tolist()
     return distance
 
 
@@ -66,7 +66,7 @@ def l2_distance(delta: np.ndarray) -> List[float]:
     List[float]
     """
     squared_difference = np.square(np.abs(delta))
-    distance = np.sum(squared_difference, axis=1, dtype=np.float).tolist()
+    distance = np.sum(squared_difference, axis=1, dtype=np.float32).tolist()
     return distance
 
 
@@ -85,7 +85,7 @@ def linf_distance(delta: np.ndarray) -> List[float]:
     """
     absolute_difference = np.abs(delta)
     # get the largest change per row
-    largest_difference = np.max(absolute_difference, axis=1)
+    largest_difference = np.max(absolute_difference, axis=1).astype(np.float32)
     distance = largest_difference.tolist()
     return distance
